pyproject.toml
[tool.poetry]
name = "laneatt-classproj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
torch = "2.5.1"
torchvision = "0.20.1"
numpy = "<2.0"
opencv-python = "<4.10"
matplotlib = "^3.9.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.0"
rich = "^13.7.0"
tensorboard = "^2.17.0"
scikit-learn = "^1.5.0"
pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
